+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleLanguage.java
@@ -578,6 +578,26 @@ public abstract class TruffleLanguage<C> {
     }
 
     /**
+     * Looks an additional language service up. By default it checks if the language itself is
+     * implementing the requested class and if so, it returns <code>this</code>.
+     * <p>
+     * In future this method can be made protected and overridable by language implementors to
+     * create more dynamic service system.
+     *
+     * @param <T> the type to request
+     * @param clazz
+     * @return
+     */
+    final /* protected */ <T> T lookup(Class<T> clazz) {
+        if (clazz.isInterface()) {
+            if (clazz.isInstance(this)) {
+                return clazz.cast(this);
+            }
+        }
+        return null;
+    }
+
+    /**
      * Find a meta-object of a value, if any. The meta-object represents a description of the
      * object, reveals it's kind and it's features. Some information that a meta-object might define
      * includes the base object's type, interface, class, methods, attributes, etc.
@@ -1223,6 +1243,11 @@ public abstract class TruffleLanguage<C> {
             return env.findSourceLocation(obj);
         }
 
+        @Override
+        public <S> S lookup(LanguageInfo language, Class<S> type) {
+            TruffleLanguage<?> spi = AccessAPI.nodesAccess().getLanguageSpi(language);
+            return spi.lookup(type);
+        }
     }
 }